[#906] Move axios interceptors to dedicated module
